Display JIRA links in own column.
This code attempts to parse the title into a metadata portion
followed by the actual title.  When possible, we remove information
from the title that's displayed in other columns (or via the tabs).

This makes it easy to quickly see which patches don't have JIRAs
and helps to declutter the table.

I also shortened some of the column labels.

Inspired by #9.