Sync JIRA data to dashboard
This pull request adds automatic synchronization of JIRA information to the dashboard.  Most of the changes are in the back-end, where I've added tasks to fetch JIRA information (using the JIRA REST API), a cron job which consumes the JIRA RSS feed in order to keep the database up-to-date, and a model to persist JIRA information into the AppEngine datastore.

On the front-end, I've added two new columns, "Priority" and "Issue Type", which show information for the first JIRA issue linked to the PR.

Here's a screenshot showing how this looks:

![image](https://cloud.githubusercontent.com/assets/50748/5384924/87d2b882-8072-11e4-86e7-48d557eb141a.png)

Closes #37.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #38 from JoshRosen/jira-integration and squashes the following commits:

cfa4edf [Josh Rosen] Undo a debugging change.
1134613 [Josh Rosen] Update queue settings.
596acc4 [Josh Rosen] Use app version as cache key prefix.
33aca1c [Josh Rosen] Bump app version so that we can do a staged rollout.
01b330f [Josh Rosen] Add JIRA bulk-loading task.
592c04f [Josh Rosen] Set 'title' attribute for JIRA icons.
20b1c3e [Josh Rosen] Display issue types + priorities in front-end.
cea8e0b [Josh Rosen] Fetch data from JIRA REST API and store in database
7a0a083 [Josh Rosen] Disable caching for configuration KVS reads
5321078 [Josh Rosen] Use watermarks to avoid re-fetching JIRA updates
4ac69f4 [Josh Rosen] Basic usage of feedparser for consuming JIRA RSS;