Replace hard-coded Spark settings with config variables
While playing around with the spark-pr-dashboard, I discovered that the variables in the settings file are not always respected. That led to a pretty nasty outcome :) Now, I would like to fix this such that other people, who blindly run the dashboard don't experience anything similar.

Other than that, kudos to everyone who helped putting this together. It's a nice piece of software.

Author: Max <max@posteo.de>

Closes #61 from mxm/config and squashes the following commits:

8a5d4ff [Max] replace hard-coded Spark settings with config variables