Added a dashboard for tracking stale PRs
Fixes #1

I added a new Stale PRs dash and sorting options for all, hanging, or abandoned PRs. These new dashes are still sortable by component as have all the features of the original dash.

Known behavioral oddity: When a component tab (such as Mesos) is selected and you you switch to a status tab (like Hanging or Abandoned) where there are no PRs in that component it will show the All components tab, but if you switch back to a status tab where the component exists it will switch back to it (instead of staying on all). This is caused by the default behavior of showing the All tab when the hash in the url matches an nonexistent tab.

Open Question: Is the UI ok? It works and doesn't look too bad, but the two sorting tab rows with the same color right on top of each other feels a bit off. What are some suggestions to fix this?

Screenshots:
![all-all](https://cloud.githubusercontent.com/assets/13952758/14869804/d11177f4-0c8a-11e6-9588-3c50ffcfb97d.png)
![all-build](https://cloud.githubusercontent.com/assets/13952758/14869817/d990cdee-0c8a-11e6-86b2-28fef50b284d.png)
![hang-all](https://cloud.githubusercontent.com/assets/13952758/14869823/dd0372f6-0c8a-11e6-86e2-d91f54db0113.png)
![hang-build](https://cloud.githubusercontent.com/assets/13952758/14869827/de5087ac-0c8a-11e6-8a75-40d32ed58b99.png)
![aband-all](https://cloud.githubusercontent.com/assets/13952758/14869829/df9fa8fe-0c8a-11e6-9cf4-e1d8b454d3c0.png)
![aband-build](https://cloud.githubusercontent.com/assets/13952758/14869831/e0d5832e-0c8a-11e6-905e-09bab98d11e7.png)

Author: Alex Bozarth <ajbozart@us.ibm.com>

Closes #71 from ajbozarth/stale-dash and squashes the following commits:

cc3aa1f [Alex Bozarth] fixed missed style issues
7bb6e0d [Alex Bozarth] Added hanging and ababandoned prs sorting
1ee7d35 [Alex Bozarth] Added stale prs tab