Added daily cron job to set deleted prs state not open
A fix for #72
I've added a daily cron job to run a new clean-prs function call. The function checks github for every pr listed as open in the datastore and if it returns a 404 it changes the pr's state to "deleted" so it won't show up in the dash. Currently there's only one pr stuck in this state, but this will protect it from happening again.

Author: Alex Bozarth <ajbozart@us.ibm.com>

Closes #73 from ajbozarth/deleted-prs and squashes the following commits:

7bb6fc6 [Alex Bozarth] Address review
859909c [Alex Bozarth] Switched to using the task queue
ae3fa4b [Alex Bozarth] Added daily cron job to set deleted prs state = deleted