Fixed word wrap so table is not wider than page
Updated the css so pr titles will wrap mid-word if a word is so long it would make the table wider than the page. It only wraps on long words and only the title column is affected (long usernames aren't wrapped).

Tested on Latest Safari, Chrome and FirefoxESR

Before and after screenshots of the pr title that was causing the issue currently:
<img width="1427" alt="screen shot 2016-11-04 at 2 31 24 pm" src="https://cloud.githubusercontent.com/assets/13952758/20023363/c2dfbe9a-a29c-11e6-9b5f-a00a7e7cc097.png">
<img width="1426" alt="screen shot 2016-11-04 at 2 30 09 pm" src="https://cloud.githubusercontent.com/assets/13952758/20023366/c5e6ae1e-a29c-11e6-94ca-36f3b0634d47.png">

Author: Alex Bozarth <ajbozart@us.ibm.com>

Closes #79 from ajbozarth/css-fix and squashes the following commits:

81b5590 [Alex Bozarth] Added min-width
b53402d [Alex Bozarth] Fixed word wrap css